Split out serverless log logic from logs command
